pyproject.toml
[project]
name = "seb-ocr"
version = "0.1.0"
description = "A script to perform OCR on historical documents using Gemini."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.7.1",
    "pillow>=10.4.0",
    "tenacity>=8.4.1",
    "google-api-core>=2.25.1",
    "google-genai>=1.21.1",
]
